middleware api refactor
