+++ broker.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-// Topic represents/identifies a particular Topic
+// Topic identifies a particular Topic on which messages can be published.
 type Topic string
 
 // String returns the string representation of the Topic.
@@ -12,7 +12,7 @@ func (t Topic) String() string {
 	return string(t)
 }
 
-// Broker pub-sub broker definition
+// Broker pub-sub broker definition.
 type Broker interface {
 	// Publish the given message onto the given topic
 	Publish(ctx context.Context, topic Topic, m interface{}) error
