+++ middleware/printer/printer.go
@@ -7,19 +7,18 @@ import (
 	"io"
 
 	"github.com/botchris/go-pubsub"
-	"github.com/botchris/go-pubsub/interceptor"
 )
 
 // PublishInterceptor intercepts each message and prints its content on the given writer in JSON format
-func PublishInterceptor(writer io.Writer) interceptor.PublishInterceptor {
-	return func(ctx context.Context, next interceptor.PublishHandler) interceptor.PublishHandler {
+func PublishInterceptor(writer io.Writer) pubsub.PublishInterceptor {
+	return func(ctx context.Context, next pubsub.PublishHandler) pubsub.PublishHandler {
 		return func(ctx context.Context, m interface{}, topic pubsub.Topic) error {
 			j, err := json.Marshal(m)
 			if err != nil {
 				return err
 			}
 
-			log := fmt.Sprintf("[broker] publishing @ %v: %s\n", topic, string(j))
+			log := fmt.Sprintf("[middleware] publishing @ %v: %s\n", topic, string(j))
 			if _, err := writer.Write([]byte(log)); err != nil {
 				return err
 			}
@@ -29,17 +28,17 @@ func PublishInterceptor(writer io.Writer) interceptor.PublishInterceptor {
 	}
 }
 
-// SubscribeInterceptor intercepts each message that is delivered to a subscribers and prints out its content on the
-// given writer in JSON format
-func SubscribeInterceptor(writer io.Writer) interceptor.SubscriberInterceptor {
-	return func(ctx context.Context, next interceptor.SubscribeMessageHandler) interceptor.SubscribeMessageHandler {
+// SubscribeInterceptor intercepts each message that is delivered to a subscribers and prints out its content on
+// the given writer in JSON format
+func SubscribeInterceptor(writer io.Writer) pubsub.SubscriberInterceptor {
+	return func(ctx context.Context, next pubsub.SubscribeMessageHandler) pubsub.SubscribeMessageHandler {
 		return func(ctx context.Context, s *pubsub.Subscriber, m interface{}) error {
 			j, err := json.Marshal(m)
 			if err != nil {
 				return err
 			}
 
-			log := fmt.Sprintf("[broker] received: %s\n", string(j))
+			log := fmt.Sprintf("[middleware] received: %s\n", string(j))
 			if _, err := writer.Write([]byte(log)); err != nil {
 				return err
 			}
